changes to work without interoperability

+++ addons/ofxGui/src/ofxColorPicker.h
@@ -19,25 +19,25 @@ template<typename ColorType>
 class ofxColorPicker_: public ofxBaseGui {
 
 public :
-	
+
 	ofxColorPicker_();
 	ofxColorPicker_(ofParameter<ofColor_<ColorType>> parameter, float w = defaultWidth, float h = defaultWidth / 9. * 16.);
 	ofxColorPicker_ * setup(ofParameter<ofColor_<ColorType>> parameter, float w = defaultWidth, float h = defaultWidth / 9. * 16.);
 
 	void setShape(float x, float y, float w, float h);
 	void setShape(ofRectangle r);
-    
+
     void setColorScale(float value);
     float getColorScale();
-    
+
     void setColorRadius(float value);
     float getColorRadius();
-    
+
     void setColorAngle(float value);
 	float getColorAngle();
 
 	ofAbstractParameter & getParameter();
-	
+
 private :
 	void render();
 	bool setValue(float mx, float my, bool bCheckBounds);
@@ -55,14 +55,14 @@ private :
 	bool mouseScrolled(ofMouseEventArgs & args);
 
 	bool mouseUpdate(ofMouseEventArgs& args);
-	
+
 	glm::vec2 colorPoint;
     float colorScale;
     float colorRadius;
 	float colorAngle;
 	bool bSettingColor = false;
 
-    int colorWheelRadius;
+    float colorWheelRadius;
     int colorWheelPad;
 
 	ofRectangle rectBackground;
